requirements.txt
argparse==1.2.1
chardet==2.2.1
cssselect==0.9.1
feedparser==5.1.3
lxml==3.3.5
readability-lxml==0.3.0.3
wsgiref==0.1.2
